refactor: extract error logic to separate hooks
